Final Grad Project
This is the project just using php script to return the value sent by
the app. Next step is to use firebase and send different messages that
receive different messages. Then, after that, I’m going to connect it
to a microcontroller so we can communicate between devices via
HTTP/wifi.